requirement.txt
1.flask
2.flask-user
3.flask-Alchemy
